+++ provisioning/chef/memcached/metadata.rb
@@ -1,9 +1,10 @@
+name              "memcached"
 maintainer        "Opscode, Inc."
 maintainer_email  "cookbooks@opscode.com"
 license           "Apache 2.0"
 description       "Installs memcached and provides a define to set up an instance of memcache via runit"
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           "1.0.2"
+version           "1.1.0"
 depends           "runit"
 
 recipe "memcached", "Installs and configures memcached"