+++ provisioning/chef/memcached/recipes/default.rb
@@ -24,7 +24,7 @@ end
 package "libmemcache-dev" do
   case node[:platform]
   when "redhat","centos","fedora"
-    package_name "libmemcache-devel"
+    package_name "libmemcached-devel"
   else
     package_name "libmemcache-dev"
   end
@@ -62,6 +62,7 @@ else
     :listen => node[:memcached][:listen],
     :user => node[:memcached][:user],
     :port => node[:memcached][:port],
+    :maxconn => node[:memcached][:maxconn],
     :memory => node[:memcached][:memory]
   )
   notifies :restart, resources(:service => "memcached"), :immediately