+++ provisioning/chef/postgresql/recipes/client.rb
@@ -19,8 +19,12 @@
 # limitations under the License.
 #
 
+if(node['postgresql']['enable_pitti_ppa'])
+  include_recipe 'postgresql::ppa_pitti_postgresql'
+end
+
 node['postgresql']['client']['packages'].each do |pg_pack|
-  package pg_pack do
-    action :install
-  end
+
+  package pg_pack
+
 end