+++ provisioning/chef/postgresql/recipes/server_redhat.rb
@@ -39,24 +39,31 @@ user "postgres" do
 end
 
 node['postgresql']['server']['packages'].each do |pg_pack|
-  package pg_pack do
-    action :install
+
+  package pg_pack
+
+end
+
+case node['platform_family']
+when "rhel"
+  case
+  when node['platform_version'].to_f >= 6.0
+    package "postgresql-server"
+  else
+    package "postgresql#{node['postgresql']['version'].split('.').join}-server"
+  end
+when "fedora", "suse"
+  package "postgresql-server"
+end
+
+if node['postgresql']['version'].to_f < 9.0
+  execute "/sbin/service #{node['postgresql']['server']['service_name']} initdb" do
+    not_if { ::FileTest.exist?(File.join(node['postgresql']['dir'], "PG_VERSION")) }
   end
 end
 
-execute "/sbin/service postgresql initdb" do
-  not_if { ::FileTest.exist?(File.join(node.postgresql.dir, "PG_VERSION")) }
-end
-
 service "postgresql" do
+  service_name node['postgresql']['server']['service_name']
   supports :restart => true, :status => true, :reload => true
   action [:enable, :start]
-end
-
-template "#{node[:postgresql][:dir]}/postgresql.conf" do
-  source "redhat.postgresql.conf.erb"
-  owner "postgres"
-  group "postgres"
-  mode 0600
-  notifies :restart, resources(:service => "postgresql"), :immediately
 end