+++ provisioning/chef/git/CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.1.0:
+
+* [COOK-1943] - Update to git 1.8.0
+
+## v1.0.2:
+
+* [COOK-1537] - add recipe for source installation
+
 ## v1.0.0:
 
 * [COOK-1152] - Add support for Mac OS X