+++ provisioning/chef/memcached/CHANGELOG.md
@@ -1,3 +1,12 @@
+## v1.1.0:
+
+* [COOK-1764] - Add Max Connections to memcached.conf and fix typos
+
+## v1.0.4:
+
+* [COOK-1192] - metadata doesn't include RH platforms (supported)
+* [COOK-1354] - dev package changed name on centos6
+
 ## v1.0.2:
 
 * [COOK-1081] - support for centos/rhel